usersTODO_aCorregir

+++ controllers/users.js
@@ -4,10 +4,8 @@ var Auth = require('../helpers/hash')
 var Users = require('../models/users')
 var FullContact = require('../helpers/fullContact')
 
-// List users
+// List users, allow unsigned user
 router.get('/', function (req, res) {
-  // TODO: use middleware
-
   Users.find().sort('name').select({ __v: 0, _id: 0 })
     .then(result => {
       res.json({ users: result })
@@ -17,7 +15,7 @@ router.get('/', function (req, res) {
     })
 })
 
-// Create a New User
+// Create a New User, allow unsigned user
 router.post('/', function (req, res) {
   if (req.body.pass.length >= 8) {
     Auth.hash(req.body.pass, hash => {
@@ -43,50 +41,53 @@ router.post('/', function (req, res) {
   }
 })
 
-// Update a user
+// Update a user, only signed user
 // TODO: actualizar una sola cosa
+// TODO: corregir!!!
 router.put('/:id', function (req, res) {
-  // let id = req.query.id // TODO: middleware
-
-  Users.findOne({ id: req.params.id }, (err, result) => {
-    if (result) {
-      Users.updateOne({ id: req.params.id }, {
-        email: req.body.email,
-        digest: req.body.digest,
-        name: req.body.name
-      }).then((err, status) => {
-        res.json({ result: 'ok' })
-      })
-        .catch(err => {
-          console.error('Something went wrong', err)
-          res.json({ message: 'Something went wrong', error: err.message })
-        })
-    } else {
-      res.json({ errors: 'user not found' })
-    }
-  })
-})
-
-// Delete a user
-router.delete('/:id', function (req, res) {
-  // let id = req.query.id // TODO: middleware
-
-  console.log(req.params.id)
-  Users.findOne({ id: req.params.id }, (err, result) => {
-    console.log(result)
-    if (result) {
-      Users.deleteOne({ id: req.params.id })
-        .then((err, status) => {
+  if (req.user) {
+    Users.findOne({ id: req.user.id }, (err, result) => {
+      if (result) {
+        Users.updateOne({ id: req.params.id }, {
+          email: req.body.email,
+          digest: req.body.digest,
+          name: req.body.name
+        }).then((err, status) => {
           res.json({ result: 'ok' })
         })
-        .catch(err => {
-          console.error('Something went wrong', err)
-          res.json({ message: 'Something went wrong', error: err })
-        })
-    } else {
-      res.json({ errors: 'user not found' })
-    }
-  })
+          .catch(err => {
+            console.error('Something went wrong', err)
+            res.json({ message: 'Something went wrong', error: err.message })
+          })
+      } else {
+        res.json({ errors: 'user not found' })
+      }
+    })
+  } else {
+    res.json({ errors: 'please sing in' })
+  }
+})
+
+// Delete a user, only signed user
+// TODO: corregir!!!
+router.delete('/:id', function (req, res) {
+  if (req.user) {
+    Users.findOne({ id: req.user.id }, (err, result) => {
+      if (result) {
+        Users.deleteOne({ id: req.params.id })
+          .then((err, status) => {
+            res.json({ result: 'ok' })
+          })
+          .catch(err => {
+            res.json({ message: 'Something went wrong', error: err })
+          })
+      } else {
+        res.json({ errors: 'user not found' })
+      }
+    })
+  } else {
+    res.json({ errors: 'please sing in' })
+  }
 })
 
 module.exports = router
